Added Website Template  Edit from Template

+++ app/Models/PageContent.php
@@ -25,6 +25,7 @@ class PageContent extends Model
         parent::boot();        
         static::deleting(function($model) {
             $model->getContent()->delete();
+            if($model->url)
             $model->url->delete();
         });
     }
